Eliminated any2stringadd.
scala> List(1, 2, 3) + "4"
<console>:8: error: value + is not a member of List[Int]
              List(1, 2, 3) + "4"
                            ^

Versus regular scala where it's

scala> List(1, 2, 3) + "4"
res0: String = List(1, 2, 3)4

The many empty traits in this commit are because for some
reason nested objects incur errors under separate compilation.
scalac is out there looking for a non-existent Foo$Bar class
when there is only a Foo$Bar$. Creating dummy traits creates the
class it's looking for, which it will never use.

Also eliminated fallbackStringCanBuildFrom, a method which
is called zillions of times in a compilation run but was needed
at exactly one call site. It was so busy because an applicable
implicit in Predef is always chosen over an applicable implicit
in a companion object, even if it's a "low priority" "fallback"
which makes no sense. Between this and contravariance selection
being backward, CanBuildFrom is a first-class snafu.

There will be some failing tests here as the test suite still
requires some de-any2stringadd-izing or less likely but possible
a bit of de-fallbackStringCanBuildFrom-izing.