Removed the "magical" String.+ method.
This was defensible when it was written, but there are better
ways to do this sort of thing now. Those ways aren't included
in this commit but perhaps I can find a volunteer. The first
thing to try I think is a fasttrack macro which flattens a series
of string concatenations into the usual StringBuilder business.

For now we'll live with a value class which is efficient when
adding two strings but less so with more. It would be nice if
we could return StringAddition from `+` instead of String as
this would solve all our performance problems, though at some
cost in sketchiness if we opted to edit the buffer in-place
assuming a "poor man's linear type". That sketchiness is moot
though because:

  val x = "a" + b + c + d

If the type of x isn't inferred as String, too much will break.
An implicit from StringAddition back down to String will only have
any effect if x is ascribed as String. I can think of a lot of ways
to make this work, but none will be possible in normal scala
because it offers no means of influencing type inference.

Fortunately we need not suffer scala's limitations.